requirements.txt
fastapi==0.115.0
mangum==0.18.0
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator>=2.1.1
uvicorn==0.32.0
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
supabase==2.10.0
google-api-python-client==2.147.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pyjwt>=2.10.1

